fix: refactor directory structure
