+++ web/src/auth/mfaForm/MfaAuthVerifyForm.js
@@ -15,9 +15,11 @@
 import React, {useState} from "react";
 import i18next from "i18next";
 import {Button, Input} from "antd";
-import * as AuthBackend from "./AuthBackend";
-import {EmailMfaType, RecoveryMfaType, SmsMfaType} from "./MfaSetupPage";
-import {MfaSmsVerifyForm, MfaTotpVerifyForm, mfaAuth} from "./MfaVerifyForm";
+import * as AuthBackend from "../AuthBackend";
+import {EmailMfaType, RecoveryMfaType, SmsMfaType} from "../MfaSetupPage";
+import {mfaAuth} from "./MfaVerifyForm";
+import MfaVerifySmsForm from "./MfaVerifySmsForm";
+import MfaVerifyTotpForm from "./MfaVerifyTotpForm";
 
 export const NextMfa = "NextMfa";
 export const RequiredMfa = "RequiredMfa";
@@ -64,29 +66,29 @@ export function MfaAuthVerifyForm({formValues, oAuthParams, mfaProps, applicatio
     return (
       <div style={{width: 300, height: 350}}>
         <div style={{marginBottom: 24, textAlign: "center", fontSize: "24px"}}>
-          {i18next.t("mfa:Multi-factor authentication")}
+          {i18next.t("mfaForm:Multi-factor authentication")}
         </div>
         <div style={{marginBottom: 24}}>
-          {i18next.t("mfa:Multi-factor authentication description")}
+          {i18next.t("mfaForm:Multi-factor authentication description")}
         </div>
         {mfaType === SmsMfaType || mfaType === EmailMfaType ? (
-          <MfaSmsVerifyForm
+          <MfaVerifySmsForm
             mfaProps={mfaProps}
             method={mfaAuth}
             onFinish={verify}
             application={application}
           />) : (
-          <MfaTotpVerifyForm
+          <MfaVerifyTotpForm
             mfaProps={mfaProps}
             onFinish={verify}
           />
         )}
         <span style={{float: "right"}}>
-          {i18next.t("mfa:Have problems?")}
+          {i18next.t("mfaForm:Have problems?")}
           <a onClick={() => {
             setMfaType("recovery");
           }}>
-            {i18next.t("mfa:Use a recovery code")}
+            {i18next.t("mfaForm:Use a recovery code")}
           </a>
         </span>
       </div>
@@ -95,12 +97,12 @@ export function MfaAuthVerifyForm({formValues, oAuthParams, mfaProps, applicatio
     return (
       <div style={{width: 300, height: 350}}>
         <div style={{marginBottom: 24, textAlign: "center", fontSize: "24px"}}>
-          {i18next.t("mfa:Multi-factor recover")}
+          {i18next.t("mfaForm:Multi-factor recover")}
         </div>
         <div style={{marginBottom: 24}}>
-          {i18next.t("mfa:Multi-factor recover description")}
+          {i18next.t("mfaForm:Multi-factor recover description")}
         </div>
-        <Input placeholder={i18next.t("mfa:Recovery code")}
+        <Input placeholder={i18next.t("mfaForm:Recovery code")}
           style={{marginBottom: 24}}
           type={"passcode"}
           size={"large"}
@@ -112,11 +114,11 @@ export function MfaAuthVerifyForm({formValues, oAuthParams, mfaProps, applicatio
           }}>{i18next.t("forget:Verify")}
         </Button>
         <span style={{float: "right"}}>
-          {i18next.t("mfa:Have problems?")}
+          {i18next.t("mfaForm:Have problems?")}
           <a onClick={() => {
             setMfaType(mfaProps.mfaType);
           }}>
-            {i18next.t("mfa:Use SMS verification code")}
+            {i18next.t("mfaForm:Use SMS verification code")}
           </a>
         </span>
       </div>
